# specs/001-a-custom-ui/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Custom Agent UI Backend API
  version: 1.0.0
paths:
  /workspaces:
    get:
      summary: List all workspaces
      responses:
        '200':
          description: A list of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workspace'
    post:
      summary: Create a new workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '201':
          description: The created workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
  /workspaces/{workspaceId}:
    get:
      summary: Get a single workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A single workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
    delete:
      summary: Delete a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content
  /workspaces/{workspaceId}/files:
    get:
      summary: List all files in a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/File'
    post:
      summary: Create a new file in a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                content:
                  type: string
      responses:
        '201':
          description: The created file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
  /files/{fileId}:
    put:
      summary: Update a file
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
      responses:
        '200':
          description: The updated file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
    delete:
      summary: Delete a file
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content
  /agent/run:
    post:
      summary: Run the agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                persona:
                  type: string
                prompt:
                  type: string
      responses:
        '200':
          description: The agent's response
          content:
            application/json:
              schema:
                type: object
                properties:
                  response:
                    type: string
components:
  schemas:
    Workspace:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        files:
          type: array
          items:
            $ref: '#/components/schemas/File'
    File:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        content:
          type: string